Implements the home page layout, getting the products from the database
