+++ src/components/App.js
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import ResetCSS from "../styles/ResetCSS";
 import "react-loader-spinner/dist/loader/css/react-spinner-loader.css";
-import Home from "./Home";
+import Home from "../components/Home/Home.jsx";
 
 export default function App() {
   return (
